requirements.txt
numpy
pandas
opencv-python
albumentations
pyyaml
lxml
tqdm
matplotlib
scikit-learn